fonts, images and js are done
