+++ src/blocks/place-card/place-card.css
@@ -9,4 +9,4 @@
   background-color: #fff;
   border-radius: 10px;
   min-height: 360px;
-} 
+}
