+++ src/scripts/Api.js
@@ -1,4 +1,4 @@
-class Api {
+export class Api {
   constructor(config) {
     this._url = config.url;
     this._headers = config.headers;
