+++ src/scripts/Card.js
@@ -1,4 +1,4 @@
-class Card {
+export class Card {
 
   constructor(name, link) {
     this.name = name;
