+++ src/style.css
@@ -1,5 +1,5 @@
 @import url(./vendor/normalize.css);
-/* @import url(./vendor/fonts.css); */
+@import url(./vendor/fonts.css);
 @import url(./blocks/root/root.css);
 @import url(./blocks/header/header.css);
 @import url(./blocks/logo/logo.css);
